agent/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
requests
google-genai
pymongo